added some try..catch for the security.

+++ demo/src/main/java/com/example/ClotheDAO.java
@@ -8,8 +8,8 @@ import lombok.AllArgsConstructor;
 @Getter
 @AllArgsConstructor
 public class ClotheDAO {
-    private String name; 
+    private String name;
     private BigDecimal price;
-    private byte quantity;
+    private short quantity;
     private String description;
 }
